+++ src/_pytest/fixtures.py
@@ -2,7 +2,6 @@
 import abc
 from collections import defaultdict
 from collections import deque
-from contextlib import suppress
 import dataclasses
 import functools
 import inspect
@@ -414,9 +413,7 @@ class FixtureRequest(abc.ABC):
             # We arrive here because of a dynamic call to
             # getfixturevalue(argname) usage which was naturally
             # not known at parsing/collection time.
-            parent = self._pyfuncitem.parent
-            assert parent is not None
-            fixturedefs = self._fixturemanager.getfixturedefs(argname, parent)
+            fixturedefs = self._fixturemanager.getfixturedefs(argname, self._pyfuncitem)
             if fixturedefs is not None:
                 self._arg2fixturedefs[argname] = fixturedefs
         # No fixtures defined with this name.
@@ -579,7 +576,6 @@ class FixtureRequest(abc.ABC):
         # (latter managed by fixturedef)
         argname = fixturedef.argname
         funcitem = self._pyfuncitem
-        scope = fixturedef._scope
         try:
             callspec = funcitem.callspec
         except AttributeError:
@@ -587,13 +583,13 @@ class FixtureRequest(abc.ABC):
         if callspec is not None and argname in callspec.params:
             param = callspec.params[argname]
             param_index = callspec.indices[argname]
-            # If a parametrize invocation set a scope it will override
-            # the static scope defined with the fixture function.
-            with suppress(KeyError):
-                scope = callspec._arg2scope[argname]
+            # The parametrize invocation scope overrides the fixture's scope.
+            scope = callspec._arg2scope[argname]
         else:
             param = NOTSET
             param_index = 0
+            scope = fixturedef._scope
+
             has_params = fixturedef.params is not None
             fixtures_not_supported = getattr(funcitem, "nofuncargs", False)
             if has_params and fixtures_not_supported: